filter

+++ src/main/java/com/gabrieljuliao/Contacts/controller/DeleteContactController.java
@@ -1,6 +1,8 @@
 package com.gabrieljuliao.Contacts.controller;
 
 import com.gabrieljuliao.Contacts.dao.ContactDAO;
+import com.gabrieljuliao.Contacts.model.Contact;
+import com.gabrieljuliao.Contacts.model.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -14,7 +16,13 @@ public class DeleteContactController extends HttpServlet {
     @Override
     protected void service(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         ContactDAO contactDAO = new ContactDAO((Connection) getServletContext().getAttribute("dbCon"));
-        contactDAO.deleteContact(req.getParameter("contact_id"));
-        resp.sendRedirect("overview");
+        User user = (User) req.getSession().getAttribute("user");
+        Contact contact = contactDAO.getContact(req.getParameter("contact_id"));
+
+        if (contact.getOwnerID().equals(user.getId())){
+            contactDAO.deleteContact(req.getParameter("contact_id"));
+            resp.sendRedirect("overview");
+        }else resp.sendError(403);
+
     }
 }
